+++ thing-list-app/src/App.js
@@ -1,4 +1,6 @@
 import React from 'react';
+import Form from 'List';
+import List from 'List';
 import './App.css';
 
 class App extends React.Component {
@@ -16,28 +18,6 @@ class App extends React.Component {
   }
 }
 
-function SnackItem(props) {
-  return <li><p>{props.snack.type}</p></li>
-}
 
-class List extends React.Component {
-  constructor(props) {
-    super(props)
-    this.state = {
-      snacks: [
-        { id: 412, type: 'Granola bar' },
-        { id: 543, type: 'Ruffles' }
-      ]
-    }
-  }
-  render() {
-    return (
-      this.state.snacks.length &&
-      <ul>
-        {this.state.snacks.map(snack => <SnackItem key={snack.id} snack={snack} />)}
-      </ul>
-    )
-  }
-}
 
 export default App;
