+++ thing-list-app/src/Footer.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+class Footer extends React.Component{
+	
+}
+
+export default Footer
