+++ thing-list-app/src/Header.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+class Header extends React.Component{
+	
+}
+
+export default Header
